+++ orm/interfaces.go
@@ -12,7 +12,6 @@ import (
 // this can be light wrapper around a protobuf-defined type
 type Object interface {
 	Keyed
-	Cloneable
 	// Validate returns error if the object is not in a valid
 	// state to save to the db (eg. field missing, out of range, ...)
 	x.Validater
@@ -30,15 +29,9 @@ type Keyed interface {
 	SetKey([]byte)
 }
 
-// Cloneable will create a new object that can be loaded into
-type Cloneable interface {
-	Clone() Object
-}
-
 // CloneableData is an intelligent Value that can be embedded
 // in a simple object to handle much of the details.
 type CloneableData interface {
 	x.Validater
 	weave.Persistent
-	Copy() CloneableData
 }