+++ orm/object.go
@@ -1,13 +1,14 @@
 package orm
 
 import (
+	"reflect"
+
 	"github.com/iov-one/weave"
 	"github.com/iov-one/weave/errors"
 	"github.com/iov-one/weave/x"
 )
 
 var _ Object = (*SimpleObj)(nil)
-var _ Cloneable = (*SimpleObj)(nil)
 var _ x.Validater = (*SimpleObj)(nil)
 
 // SimpleObj wraps a key and a value together
@@ -54,8 +55,9 @@ func (o *SimpleObj) SetKey(key []byte) {
 
 // Clone will make a copy of this object
 func (o *SimpleObj) Clone() Object {
+	cpy := reflect.New(reflect.TypeOf(o.value).Elem()).Interface().(CloneableData)
 	res := &SimpleObj{
-		value: o.value.Copy(),
+		value: cpy,
 	}
 	// only copy key if non-nil
 	if len(o.key) > 0 {