+++ x/msgfee/model.go
@@ -43,7 +43,7 @@ type MsgFeeBucket struct {
 // NewMsgFeeBucket returns a bucket for keeping track of fees for each message
 // type. Message fees are indexed by the corresponding message path.
 func NewMsgFeeBucket() *MsgFeeBucket {
-	b := migration.NewBucket("msgfee", "msgfee", orm.NewSimpleObj(nil, &MsgFee{}))
+	b := migration.NewBucket("msgfee", "msgfee", &MsgFee{})
 	return &MsgFeeBucket{
 		Bucket: b,
 	}